+++ db/migrate/20160820110411_change_cuit_in_users.rb
@@ -1,6 +1,6 @@
 class ChangeCuitInUsers < ActiveRecord::Migration[5.0]
   def up
-    change_column :users, :cuit, :double
+    change_column :users, :cuit, :float
   end
 
   def down
